[Decomp][SpellMgr] Cohesion split: extract spell-area loaders to SpellMgrAreas.cpp
